+++ messaging-ui/src/main/java/com/nabla/sdk/messaging/ui/scene/messages/adapter/viewholders/ProviderMessageViewHolder.kt
@@ -4,7 +4,6 @@ import android.view.Gravity
 import android.view.View
 import android.view.View.INVISIBLE
 import android.view.View.VISIBLE
-import androidx.annotation.VisibleForTesting
 import androidx.appcompat.widget.PopupMenu
 import androidx.core.view.isVisible
 import com.benasher44.uuid.Uuid
@@ -34,15 +33,6 @@ internal sealed class ProviderMessageViewHolder<ContentType : TimelineItem.Messa
         menuInflater.inflate(R.menu.nabla_message_actions, menu)
     }
 
-    private var avatarRandomSeedOverridden = false
-    private var overriddenAvatarRandomSeed: Any? = null
-
-    @VisibleForTesting
-    fun overrideAvatarBackgroundRandomSeed(seed: Any?) {
-        avatarRandomSeedOverridden = true
-        overriddenAvatarRandomSeed = seed
-    }
-
     override fun bind(message: TimelineItem.Message, author: TimelineItem.Message.Author.Provider, content: ContentType) {
         super.bind(message, author, content)
 
@@ -55,12 +45,7 @@ internal sealed class ProviderMessageViewHolder<ContentType : TimelineItem.Messa
         }
 
         if (message.showAuthorAvatar) {
-            if (avatarRandomSeedOverridden) {
-                binding.chatProviderMessageAvatarView.loadAvatar(author.provider, overriddenAvatarRandomSeed)
-            } else {
-                binding.chatProviderMessageAvatarView.loadAvatar(author.provider)
-            }
-
+            binding.chatProviderMessageAvatarView.loadAvatar(author.provider)
             binding.chatProviderMessageAvatarViewContainer.setOnClickListener { author.provider.id.let(onProviderClicked) }
         }
     }