+++ messaging-ui/src/test/java/com/nabla/sdk/messaging/ui/messages/ConversationOtherTextCellTest.kt
@@ -250,7 +250,6 @@ internal class ConversationOtherTextCellTest : BaseCoroutineTest() {
             avatar = null,
         )
 
-        overrideAvatarBackgroundRandomSeed(null)
         bind(
             TimelineItem.Message.fake(
                 sendStatus = SendStatus.Sent,